refactor: consolidate 10 facade/barrel files into 2
- 6 DOM facades (workspace-dom, settings-dom, file-dom, tab-dom,
  terminal-dom, git-dom) merged into src/utils/dom-facades.js
- 4 manager barrels (data-managers, infra-managers, io-managers,
  workflow-managers) merged into main/managers.js
- 4 subsystem facades left unchanged (legitimate grouping purpose)
- All imports updated across 33 consumer files
- Build and tests pass (398/398)

Closes #462
