+++ src/utils/file-tree-renderer.js
@@ -4,7 +4,7 @@
  */
 
 import { emitFileOpen } from './workspace-events.js';
-import { _el, createActionButton } from './file-dom.js';
+import { _el, createActionButton } from './dom-facades.js';
 import { buildChevronRow } from './chevron-row.js';
 import { computeIndent, CHEVRON_EXPANDED, CHEVRON_COLLAPSED, SVG_ICONS, HEADER_ACTIONS } from './file-tree-helpers.js';
 import { buildFileContextItems, buildDirContextItems } from './file-tree-context-menu.js';
