+++ src/utils/file-viewer-mode-bar.js
@@ -3,7 +3,7 @@
  * Extracted from file-viewer.js to reduce component size.
  */
 
-import { _el, renderList } from './file-dom.js';
+import { _el, renderList } from './dom-facades.js';
 import { STATIC_MODES } from './editor-helpers.js';
 
 /**
