--- src/utils/tab-dom.js
@@ -1,8 +1,0 @@
-/**
- * DOM re-exports for the tab domain.
- *
- * Tab-related modules (tab-bar-renderer, tab-renderer, tab-color-filter,
- * tab-lifecycle) import _el through this facade instead of reaching into
- * the core dom.js hub directly.  This reduces dom.js fan-in.
- */
-export { _el, renderList } from './dom.js';
